Move ColorPanel types into their own types.ts file alongside the ColorPanel component
